chore/destination_elder: extract out IDataHandler
